Update GeminiSelection and create GeminiChat panel
Move main chat panel to it's own component
